refactor(customer-portal): Update import paths for useAuth hook in Checkout, Dashboard, and MyOrders components
