+++ apps/factory-portal/src/App.jsx
@@ -4,56 +4,6 @@ import { AuthProvider } from './contexts/AuthContext';
 import { ProtectedRoute, AdminRoute } from './components/ProtectedRoute';
 import Login from './pages/Login';
 import Dashboard from './pages/Dashboard';
-
-// Create a theme with factory/admin colors
-const theme = createTheme({
-  palette: {
-    primary: {
-      main: '#1e88e5', // Blue shade
-    },
-    secondary: {
-      main: '#ff3d00', // Orange shade
-    },
-  },
-});
-
-function App() {
-  return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <AuthProvider>
-        <Router>
-          <Routes>
-            {/* Public routes */}
-            <Route path="/login" element={<Login />} />
-
-            {/* Protected routes for both admin and subadmin */}
-            <Route element={<ProtectedRoute />}>
-              <Route path="/dashboard" element={<Dashboard />} />
-              {/* Add more protected routes here that both roles can access */}
-            </Route>
-
-            {/* Admin-only routes */}
-            <Route element={<AdminRoute />}>
-              {/* Add admin-only routes here */}
-            </Route>
-
-            {/* Redirect to login if no route matches */}
-            <Route path="*" element={<Navigate to="/login" replace />} />
-          </Routes>
-        </Router>
-      </AuthProvider>
-    </ThemeProvider>
-  );
-}
-
-export default App;
-import { BrowserRouter as Router, Routes, Route, Navigate } from 'react-router-dom';
-import { CssBaseline, ThemeProvider, createTheme } from '@mui/material';
-import { AuthProvider } from './contexts/AuthContext';
-import { ProtectedRoute, AdminRoute } from './components/ProtectedRoute';
-import Login from './pages/Login';
-import Dashboard from './pages/Dashboard';
 import Inventory from './pages/Inventory';
 import InventoryDetail from './pages/InventoryDetail';
 import Orders from './pages/Orders';
